Add file tree preview
